Refactor: Extracted MrLavaLoba method
The MrLavaLoba method (by de'Michieli et al.) is now in a separate
header file. Simulation class had to be modified.
